# docs/swagger.yaml
basePath: /
definitions:
...
  handler.UpdateProfileRequest:
    properties:
      avatar_url:
        type: string
      display_name:
        maxLength: 100
        type: string
    type: object
  handler.UpdateSettingsRequest:
    properties:
      language:
        maxLength: 10
        type: string
      notifications_enabled:
        type: boolean
      theme:
        maxLength: 20
        type: string
      timezone:
        maxLength: 50
        type: string
    type: object
...
  response.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrorDetail'
      success:
        type: boolean
    type: object
  response.SuccessResponse:
    properties:
      data: {}
      meta: {}
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Saybridge Platform Backend API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Saybridge API
  version: "1.0"
paths:
...
  /api/v1/search/messages:
    get:
      description: Full-text search across messages in accessible rooms
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - description: Filter by room ID
        in: query
        name: room_id
        type: string
      - default: 20
        description: Max results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Search results
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Search failed
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search messages
      tags:
      - Search
  /api/v1/search/rooms:
    get:
      description: Search for rooms by name within accessible rooms for the user
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - default: 20
        description: Max results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Matching rooms
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Search failed
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search rooms
      tags:
      - Search
  /api/v1/search/users:
    get:
      description: Search for users by username or display name within the tenant
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - default: 20
        description: Max results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Matching users
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Search failed
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search users
      tags:
      - Search
  /api/v1/users/me:
    get:
      description: Returns the authenticated user's profile and settings
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Update the authenticated user's display name and/or avatar URL
      parameters:
      - description: Profile updates
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated profile
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Update failed
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - Users
  /api/v1/users/me/settings:
    patch:
      consumes:
      - application/json
      description: Update user preferences such as language, theme, timezone, and
        notification settings
      parameters:
      - description: Settings updates
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated settings
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Update failed
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user settings
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    description: Enter your JWT token with the Bearer prefix, e.g. "Bearer eyJhbGci..."
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"